Simplify copy and create_file functions.
We use to call mktemp -d inside the schroot, however this is not needed.
It also makes assumptions that we can connect to the schroot as the
current user, when the caller requested we connect with the user
specified in the user= argument.

Instead, we can create the temporary file outside the schroot. We do
this with the tempfile module, part of Python core. This makes the code
simpler and easier to understand.

Also use "cp" instead of "mv" inside the chroot, this means we get new
default permissions for the file, not the permissions that were the temp
file had, which probably aren't appropriate.